+++ src/game/client/tf/tf_hud_mainmenuoverride.cpp
@@ -190,6 +190,11 @@ CHudMainMenuOverride::CHudMainMenuOverride( IViewPort *pViewPort ) : BaseClass( 
 	m_flCheckTrainingAt = 0;
 	m_bWasInTraining = false;
 
+	if ( g_pVGuiLocalize )
+	{
+		g_pVGuiLocalize->AddFile( "resource/tfsolo_%language%.txt" );
+	}
+
 	ScheduleItemCheck();
 
  	m_pToolTip = new CMainMenuToolTip( this );